Updates script
